Copy and re-push value when setlocal and setinstancevariable
We used to pop value when setting variables, but now we keep the value
on the stack and let compiler decide if the value should be popped.

By default assign expression should be treated as statement, which means
IsStmt should be true and compiler will add 'pop' instruction after
setting variable. But if its IsStmt is setted to false, compiler won't
put pop after assign instruction.