Change builtin method failure test's expected CFP.
Let's say we have following code:

```ruby
['T', 'A', 'I', 'P', 'E', 'I'].each(101) do |char|
  puts char
end
```

In the past the method will fail with CFP is 2. Because there'll be the
base call frame and the method call's block frame. We used to keep the
block frame because the program will be stopped when error occurs. So the frame won't be
popped.

But now even if an error occurs, the *thread.evalCallFrame() function
will still cleanup useless block frames as normal condition.