Fix method call parsing issue
In some cases range type might be parsed as a method call's argument.
For example:

```ruby
bar.foo
(1..5).to_a
```
In this case the range object `(1..5)` would be parsed as method `#foo`'s argument
instead of a range type. This is because parseCallExpressionWithReceiver
function didn't check if the next token is at same line and looking for
arguments directly.